airdrop/requirements.txt
python-telegram-bot==20.7
fastapi==0.109.0
uvicorn==0.27.0
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
python-dotenv==1.0.0
httpx==0.25.2
Pillow==10.2.0
alembic==1.13.1
